Refactor TUI
